\name{Rnanoflann-package}
\alias{Rnanoflann-package}
\alias{Rnanoflann}
\docType{package}
\title{
\packageTitle{Rnanoflann}
}
\description{
\packageDescription{Rnanoflann}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rnanoflann}
\packageIndices{Rnanoflann}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Rnanoflann}

Maintainer: \packageMaintainer{Rnanoflann}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
