\name{Rnanoflann-package}
\alias{Rnanoflann-package}
\alias{Rnanoflann}
\docType{package}
\title{
\packageTitle{Rnanoflann}
}
\description{
\packageDescription{Rnanoflann}
}
\details{
The package offers a function to find the k nearest neighbours for every point in a given dataset using 
Jose Luis' nanoflann library. There is support for exact searches, fixed radius searches with 'kd' trees 
and currently it supports only two distances: the 'Euclidean' and 'Manhattan'. 
Also, the nanoflann library is exported and ready ot use via the LinkingTo mehcanism.

}
\author{
\packageAuthor{Rnanoflann}

Maintainer: \packageMaintainer{Rnanoflann}
}

