\name{Rnanoflann-package}
\alias{Rnanoflann-package}
\alias{Rnanoflann}
\docType{package}
\title{
\packageTitle{Rnanoflann}
}
\description{
\packageDescription{Rnanoflann}
}
\details{
\tabular{ll}{
Package: \tab \packageName{Rnanoflann}\cr
Type: \tab \packageType{Rnanoflann}\cr
Version: \tab \packageVersion{Rnanoflann} \cr
Date: \tab \packageDate{Rnanoflann}\cr
License: \tab \packageLicense{Rnanoflann}\cr
}
}
\author{
Authors: \packageAuthors{Rnanoflann}

Maintainer: \packageMaintainer{Rnanoflann}
}
\keyword{ package }
\keyword{ Nearest Neighbors }
